test: use fixture data for overlay boundary
