+++ cmd/xgo/version.go
@@ -7,8 +7,8 @@ import "fmt"
 // VERSION is manually updated when needed a new tag
 // if you did not install git hooks, you can manually update them
 const VERSION = "1.2.3"
-const REVISION = "5babd9fcb9523c092ebdfbd23d5c4963688765ce+1"
-const NUMBER = 656
+const REVISION = "1fd3bd81685b2a37577fed0310d6d87b0273b497+1"
+const NUMBER = 657
 
 // Rationale: xgo consists of these modules:
 //
