+++ test/integrations/overlay_difference_between_go_1_19_through_1_23_vs_go_1_24_and_later/testdata/module/go.mod
@@ -1,0 +1,3 @@
+module example.com/native-overlay-vet
+
+go 1.19
