+++ test/integrations/overlay_difference_between_go_1_19_through_1_23_vs_go_1_24_and_later/testdata/module/replacement/subject.go
@@ -1,0 +1,3 @@
+package target
+
+func Value() string { return "caller replacement" }
