+++ test/integrations/overlay_difference_between_go_1_19_through_1_23_vs_go_1_24_and_later/testdata/module/subject.go
@@ -1,0 +1,7 @@
+package target
+
+import _ "example.com/native-overlay-vet/not_present"
+
+func Value() string {
+	return "original"
+}
