+++ test/integrations/overlay_difference_between_go_1_19_through_1_23_vs_go_1_24_and_later/testdata/module/subject_test.go
@@ -1,0 +1,9 @@
+package target
+
+import "testing"
+
+func TestValue(t *testing.T) {
+	if got, want := Value(), "caller replacement"; got != want {
+		t.Fatalf("Value() = %q, want %q", got, want)
+	}
+}
